Add files via upload
Since accuracy rate does not increase, most probable reason for it seems to be uneven distribution of the classes in the training set. Therefore, I plotted the error of the model : which classes it confuses most. Now, to increase the accuracy rate, i need to deal with the class imbalances.